requirements.txt
Werkzeug==0.15.2
datadog==0.28.0
google-cloud-bigquery==1.11.2
google-cloud-pubsub==0.40.0
raven==6.10.0
geoip2==2.9.0
ua-parser==0.8.0
